lab678/requirements.txt
python-dotenv>=1.0.1
requests>=2.31.0
psycopg2-binary>=2.9.9
pgvector>=0.3.6
sentence-transformers>=2.2.0
streamlit>=1.35.0
pandas>=2.2.0
